# openapi.json
{"openapi":"3.1.0","info":{"title":"Employee Management API","description":"REST API for Employee Management. Built for CI/CD demonstration.","version":"1.0.0"},"paths":{"/api/v1/health":{"get":{"tags":["Health"],"summary":"Liveness probe","description":"Basic health check endpoint to verify the application is running.","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/v1/ready":{"get":{"tags":["Health"],"summary":"Readiness probe","description":"Readiness check endpoint that verifies connectivity to the database.","operationId":"readiness_check_api_v1_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}},"503":{"description":"Service is not ready (Database connection failed)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/v1/departments":{"post":{"tags":["Departments"],"summary":"Create a new department","operationId":"create_department_api_v1_departments_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Departments"],"summary":"List departments","operationId":"list_departments_api_v1_departments_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","maximum":100000,"minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_DepartmentResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/departments/{department_id}":{"get":{"tags":["Departments"],"summary":"Get a department by ID","operationId":"get_department_api_v1_departments__department_id__get","parameters":[{"name":"department_id","in":"path","required":true,"schema":{"type":"integer","maximum":2147483647,"minimum":1,"description":"Department ID (32-bit integer)","title":"Department Id"},"description":"Department ID (32-bit integer)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Departments"],"summary":"Update a department","operationId":"update_department_api_v1_departments__department_id__put","parameters":[{"name":"department_id","in":"path","required":true,"schema":{"type":"integer","maximum":2147483647,"minimum":1,"description":"Department ID (32-bit integer)","title":"Department Id"},"description":"Department ID (32-bit integer)"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartmentResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Departments"],"summary":"Delete a department","operationId":"delete_department_api_v1_departments__department_id__delete","parameters":[{"name":"department_id","in":"path","required":true,"schema":{"type":"integer","maximum":2147483647,"minimum":1,"description":"Department ID (32-bit integer)","title":"Department Id"},"description":"Department ID (32-bit integer)"}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/departments/{department_id}/employees":{"get":{"tags":["Departments"],"summary":"List employees in a department","operationId":"list_department_employees_api_v1_departments__department_id__employees_get","parameters":[{"name":"department_id","in":"path","required":true,"schema":{"type":"integer","maximum":2147483647,"minimum":1,"description":"Department ID (32-bit integer)","title":"Department Id"},"description":"Department ID (32-bit integer)"},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","maximum":100000,"minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_EmployeeResponse_"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/employees":{"post":{"tags":["Employees"],"summary":"Create a new employee","operationId":"create_employee_api_v1_employees_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Department not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Employees"],"summary":"List all employees","operationId":"list_employees_api_v1_employees_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","maximum":100000,"minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_EmployeeResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/employees/search":{"get":{"tags":["Employees"],"summary":"Search employees by name, email, or code","operationId":"search_employees_api_v1_employees_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","minLength":2,"pattern":"^[^\\x00]*$","description":"Search term","title":"Q"},"description":"Search term"},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","maximum":100000,"minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_EmployeeResponse_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/employees/{employee_id}":{"get":{"tags":["Employees"],"summary":"Get an employee by ID","operationId":"get_employee_api_v1_employees__employee_id__get","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","maximum":2147483647,"minimum":1,"description":"Employee ID (32-bit integer)","title":"Employee Id"},"description":"Employee ID (32-bit integer)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Employees"],"summary":"Update an employee","operationId":"update_employee_api_v1_employees__employee_id__put","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","maximum":2147483647,"minimum":1,"description":"Employee ID (32-bit integer)","title":"Employee Id"},"description":"Employee ID (32-bit integer)"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Employees"],"summary":"Delete an employee","operationId":"delete_employee_api_v1_employees__employee_id__delete","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","maximum":2147483647,"minimum":1,"description":"Employee ID (32-bit integer)","title":"Employee Id"},"description":"Employee ID (32-bit integer)"}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/employees/{employee_id}/status":{"patch":{"tags":["Employees"],"summary":"Update employee active status","operationId":"update_employee_status_api_v1_employees__employee_id__status_patch","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","maximum":2147483647,"minimum":1,"description":"Employee ID (32-bit integer)","title":"Employee Id"},"description":"Employee ID (32-bit integer)"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeStatusUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"DepartmentCreate":{"properties":{"name":{"type":"string","maxLength":100,"minLength":2,"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["name"],"title":"DepartmentCreate"},"DepartmentResponse":{"properties":{"name":{"type":"string","maxLength":100,"minLength":2,"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"}},"type":"object","required":["name","id","created_at","updated_at"],"title":"DepartmentResponse"},"DepartmentUpdate":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":100,"minLength":2},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","title":"DepartmentUpdate"},"EmployeeCreate":{"properties":{"first_name":{"type":"string","maxLength":100,"minLength":2,"title":"First Name"},"last_name":{"type":"string","maxLength":100,"minLength":2,"title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"phone":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Phone"},"job_title":{"type":"string","maxLength":100,"minLength":2,"title":"Job Title"},"salary":{"type":"number","format":"decimal","title":"Salary"},"department_id":{"type":"integer","maximum":2147483647.0,"minimum":1.0,"title":"Department Id"},"is_active":{"type":"boolean","title":"Is Active","default":true}},"type":"object","required":["first_name","last_name","email","job_title","salary","department_id"],"title":"EmployeeCreate"},"EmployeeResponse":{"properties":{"first_name":{"type":"string","maxLength":100,"minLength":2,"title":"First Name"},"last_name":{"type":"string","maxLength":100,"minLength":2,"title":"Last Name"},"email":{"type":"string","format":"email","title":"Email"},"phone":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Phone"},"job_title":{"type":"string","maxLength":100,"minLength":2,"title":"Job Title"},"salary":{"type":"number","format":"decimal","title":"Salary"},"department_id":{"type":"integer","maximum":2147483647.0,"minimum":1.0,"title":"Department Id"},"is_active":{"type":"boolean","title":"Is Active","default":true},"id":{"type":"integer","title":"Id"},"employee_code":{"type":"string","title":"Employee Code"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"}},"type":"object","required":["first_name","last_name","email","job_title","salary","department_id","id","employee_code","created_at","updated_at"],"title":"EmployeeResponse"},"EmployeeStatusUpdate":{"properties":{"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["is_active"],"title":"EmployeeStatusUpdate"},"EmployeeUpdate":{"properties":{"first_name":{"anyOf":[{"type":"string","maxLength":100,"minLength":2},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string","maxLength":100,"minLength":2},{"type":"null"}],"title":"Last Name"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"phone":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Phone"},"job_title":{"anyOf":[{"type":"string","maxLength":100,"minLength":2},{"type":"null"}],"title":"Job Title"},"salary":{"anyOf":[{"type":"number","format":"decimal"},{"type":"null"}],"title":"Salary"},"department_id":{"anyOf":[{"type":"integer","maximum":2147483647.0,"minimum":1.0},{"type":"null"}],"title":"Department Id"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"}},"type":"object","title":"EmployeeUpdate"},"ErrorResponse":{"properties":{"detail":{"type":"string","title":"Detail"}},"type":"object","required":["detail"],"title":"ErrorResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"service":{"type":"string","title":"Service"}},"type":"object","required":["status","service"],"title":"HealthResponse"},"PaginatedResponse_DepartmentResponse_":{"properties":{"total":{"type":"integer","title":"Total"},"items":{"items":{"$ref":"#/components/schemas/DepartmentResponse"},"type":"array","title":"Items"},"skip":{"type":"integer","title":"Skip"},"limit":{"type":"integer","title":"Limit"}},"type":"object","required":["total","items","skip","limit"],"title":"PaginatedResponse[DepartmentResponse]"},"PaginatedResponse_EmployeeResponse_":{"properties":{"total":{"type":"integer","title":"Total"},"items":{"items":{"$ref":"#/components/schemas/EmployeeResponse"},"type":"array","title":"Items"},"skip":{"type":"integer","title":"Skip"},"limit":{"type":"integer","title":"Limit"}},"type":"object","required":["total","items","skip","limit"],"title":"PaginatedResponse[EmployeeResponse]"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}